Module Traced_sigs.Map

module type S = sig ... end

In Lwtreslib, like in the Stdlib, the Map module exports a functor to instantiate maps with known-type keys. As a result, the bulk of the documentation for maps is located within the module types returned by the functors: in Map.