Module type Traced_sigs__Set.S

In Lwtreslib, like in the Stdlib, the Set module exports mainly functors to instantiate sets with known-type keys. As a result, the bulk of the documentation for sets is located within the module types returned by the functors: in Set.

type 'error trace

'error trace is intended to be substituted by a type provided by a Trace module (with type 'error trace := 'error Trace.trace)

module type S = Traced_functor_outputs.Set.S with type 'error trace := 'error trace

module Make : functor (Ord : Stdlib.Set.OrderedType) -> S with type elt = Ord.t