Module Traced_sigs__Trace

Traces are used when errors need to be composed. This is useful directly to the user: building traces of errors to keep track of successive failures. It is also useful to the parallel traversors of this library (_ep) to combine errors that were constructed concurrently.