Module Traced_structs.Structs

module Make : functor (Trace : Traced_sigs.Trace.S) -> sig ... end

Make is a functor that takes a Trace as argument and instantiates all the Traced modules based on it.