Module Traces

module type S = Traced_sigs.Trace.S

module type EXTENDED = sig ... end

module SingletonR : EXTENDED

Singleton is a trace implementation where a trace carries exactly one error. Additional information is discarded. The different variant discard additional information in different unspecified way. This is useful for testing purpose if you need to check that you do not depend on some unspecified behaviour. SingletonND is even non-deterministic.

module SingletonL : EXTENDED

module SingletonND : EXTENDED

module Flat : EXTENDED

Flat is a trace implementation where a trace carries a flat collection of errors. No error is discarded, but the structure (parallel vs sequential) is.

module Full : EXTENDED

Full is a trace implementation where a trace carries a structured collection of errors. No error is discarded, nor is the structure.