Module type Hashtbl.S

In Lwtreslib, like in the Stdlib, the Hashtbl module exports mainly functors to instantiate hashtables with known-type keys. As a result, the bulk of the documentation for hashtables is located within the module types returned by the functors: in Hashtbl.

Note the presence of Make_es which deviates from the Stdlib to provide specialised convenience for tables of elements the initialisation of which may take time and may fail.

val hash : 'a -> int

val seeded_hash : int -> 'a -> int

val hash_param : meaningful:int -> total:int -> 'a -> int

val seeded_hash_param : meaningful:int -> total:int -> int -> 'a -> int

module type S = Bare_functor_outputs.Hashtbl.S

module Make : functor (H : Stdlib.Hashtbl.HashedType) -> S with type key = H.t

module type SeededS = Bare_functor_outputs.Hashtbl.SeededS

module MakeSeeded : functor (H : Stdlib.Hashtbl.SeededHashedType) -> SeededS with type key = H.t

module type S_ES = Bare_functor_outputs.Hashtbl.S_ES

module Make_es : functor (H : Stdlib.Hashtbl.HashedType) -> S_ES with type key = H.t