Module Lwtreslib.Bare

Instance: Bare

Bare provides all the functions as described above. It is intended to be opened to shadow some modules of Stdlib.

All values within the modules follow the same naming and semantic conventions described above. The sequential traversors are fail-early: in the following example the code returns an Error and does not print anything.

List.iter_e
   (fun x ->
      if x = "" then
         Error "empty string"
      else begin
         print_endline x;
         Ok ())
   [
      ""; (* This will cause the iteration to stop *)
      "this is not printed";
      "neither is this printed";
   ]

The concurrent (parallel) traversors are best-effort: in the following example the code prints all the non-empty strings in an unspecified order before returning an Error.

List.iter_ep
   (fun x ->
      if x = "" then
         Lwt.return (Error "empty string")
      else begin
         print_endline x;
         Lwt.return (Ok ()))
   [
      ""; (* This will cause the iteration to error in the end *)
      "this is printed";
      "this is printed as well";
   ]

The module WithExceptions provides some exception-raising helpers to reduce the boilerplate that the library imposes.

module Hashtbl : Bare_sigs.Hashtbl.S

module List : Bare_sigs.List.S

module Map : Bare_sigs.Map.S

module Monad : Bare_sigs.Monad.S

lwt monad

module Option : Bare_sigs.Option.S

module Result : Bare_sigs.Result.S

module Seq : Bare_sigs.Seq.S

module Set : Bare_sigs.Set.S

module WithExceptions : Bare_sigs.WithExceptions.S