Module Traced_sigs.Set

module type S = sig ... end

In Lwtreslib, like in the Stdlib, the Set module exports mainly functors to instantiate sets with known-type keys. As a result, the bulk of the documentation for sets is located within the module types returned by the functors: in Traced_functor_outputs.Set.